dom/html/test/forms/test_input_radio_required.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id={BUGNUMBER}
-->
<head>
  <title>Test for Bug 1100535</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1100535">Mozilla Bug 1100535</a>
<p id="display"></p>
<div id="content" style="display: none">
<form>
  <input type="radio" name="a">
</form>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
  var input = document.querySelector("input");
  input.setAttribute("required", "x");
  input.setAttribute("required", "y");
  is(document.forms[0].checkValidity(), false);
  input.required = false;
  is(document.forms[0].checkValidity(), true);
</script>
</pre>
</body>
</html>
